requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asyncmy==0.2.10
click==8.1.8
ecdsa==0.19.0
fastapi==0.115.11
h11==0.14.0
idna==3.10
passlib==1.7.4
pyasn1==0.4.8
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
python-dotenv==1.0.1
python-jose==3.4.0
rsa==4.9
six==1.17.0
sniffio==1.3.1
starlette==0.46.0
typing_extensions==4.12.2
uvicorn==0.34.0
